EhCache 2.8.1
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@2928 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core/owsi-core-components/owsi-core-component-wicket-more/src/main/java/fr/openwide/core/wicket/more/console/maintenance/ehcache/component/EhCacheCacheModificationPanel.java
@@ -31,7 +31,7 @@ public class EhCacheCacheModificationPanel extends AbstractAjaxModalPopupPanel<E
 	
 	private Form<Cache> cacheForm;
 	
-	private TextField<Integer> maxSizeField;
+	private TextField<Long> maxSizeField;
 	
 	private Component parent;
 	
@@ -51,7 +51,7 @@ public class EhCacheCacheModificationPanel extends AbstractAjaxModalPopupPanel<E
 		DelegatedMarkupPanel body = new DelegatedMarkupPanel(wicketId, EhCacheCacheModificationPanel.class);
 		
 		cacheForm = new Form<Cache>("cacheForm");
-		maxSizeField = new TextField<Integer>("maxSize", BindingModel.of(getModel(),
+		maxSizeField = new TextField<Long>("maxSize", BindingModel.of(getModel(),
 				CoreWicketMoreBinding.ehCacheCacheInformation().maxElementsInMemory()));
 		maxSizeField.setLabel(new ResourceModel("console.maintenance.ehcache.portfolio.max"));
 		cacheForm.add(maxSizeField);
@@ -71,7 +71,7 @@ public class EhCacheCacheModificationPanel extends AbstractAjaxModalPopupPanel<E
 			@Override
 			protected void onSubmit(AjaxRequestTarget target, Form<?> form) {
 				try {
-					Integer maxSize = maxSizeField.getModelObject();
+					Long maxSize = maxSizeField.getModelObject();
 					EhCacheCacheModificationPanel.this.getModelObject().setMaxElementsInMemory(maxSize);
 					Session.get().success(getString("console.maintenance.ehcache.cacheManager.cache.message.ok"));
 					closePopup(target);
